WIP - This week starting to work
a bug where if you select "this week",
and then select the project you had before selecting
it, you can't. This is likely because when selecting
"this week", the useState variable selectedProject doesn't
update, and when selecting it back, it doesn't update back.
So selecting the project back doesn't cause a rerender.